Create File for Handler and State #13 - Move Handler and State from page.tsx to handler.ts - Remove unused variable - Fix issue button Clear and Copy
