+++ src/app/globals.css
@@ -7,7 +7,7 @@
   --color-background: var(--background);
   --color-foreground: var(--foreground);
   --font-sans: "Roboto Flex", sans-serif;
-  --font-mono: "Roboto Mono", monospace;
+  --font-mono: Consolas, "Courier New", monospace;
   --font-weight: 400;
   --font-style: normal;
   --color-sidebar-ring: var(--sidebar-ring);
