+++ src/app/layout.tsx
@@ -1,7 +1,7 @@
 //- src/app/layout.tsx
 
 import type { Metadata } from "next";
-import { Roboto_Flex, Roboto_Mono } from "next/font/google";
+import { Roboto_Flex } from "next/font/google";
 import "./globals.css";
 import ReactQueryProvider from "./provider";
 import { SidebarInset, SidebarProvider } from "@/components/ui/sidebar"
@@ -10,11 +10,6 @@ import { AppFooter } from "@/components/core/app-layout"
 
 const fontSans = Roboto_Flex({
   variable: "--font-roboto-flex",
-  subsets: ["latin"],
-})
-
-const fontMono = Roboto_Mono({
-  variable: "--font-roboto-mono",
   subsets: ["latin"],
 })
 
@@ -30,7 +25,7 @@ export default function RootLayout({
 }>) {
   return (
     <html lang="en">
-      <body className={`${fontSans.variable} ${fontMono.variable} antialiased`}>
+      <body className={`${fontSans.variable} antialiased`}>
         <ReactQueryProvider>
           <SidebarProvider>
             <AppSidebar className="shadow-xs" />
